requirements.txt
    #------------------------------------------------------------------------------
#--- !!!   IMPORTANT   !!!
#    This is a Collection of all Project Dependencies.
#    !!! DO NOT REMOVE !!!
#------------------------------------------------------------------------------
#--- Upgrade all
#    pip freeze --local | grep -v "^\-e" | cut -d = -f 1 | xargs -n1 pip install -U

#------------------------------------------------------------------------------
# --- "#"
git+ssh://git@github.com/asuvorov/ddaemon-core-python.git@dev
git+ssh://git@github.com/asuvorov/ddaemon-utils-python.git@dev

Pillow==5.3.0

# -----------------------------------------------------------------------------
# --- "B"
bumpversion==0.5.3

# -----------------------------------------------------------------------------
# --- "C"
csscompressor==0.9.5

# -----------------------------------------------------------------------------
# --- "D"
djangorestframework==3.9.0

# -----------------------------------------------------------------------------
# --- "django-*"
django-bower==5.2.0
django-compressor==2.2
django-storages==1.7.1
django-twitter-tag==1.2.1

#------------------------------------------------------------------------------
# "M"
mysql-python==1.2.5

# -----------------------------------------------------------------------------
# --- "python-*"
python-decouple==3.1

# -----------------------------------------------------------------------------
# --- "S"
slimit==0.8.1

#------------------------------------------------------------------------------
# --- "T"
termcolor==1.1.0

#------------------------------------------------------------------------------
# --- "W"
whitenoise==4.1
